[chore🏗] : coupon type 정률 할인률 -> %단위이기 때문에 data type을 integer로 수정 (해당 부분 코드, 테스트 코드 수정)

+++ src/main/java/shop/gaship/coupon/coupontype/dto/CouponTypeCreationRequestDto.java
@@ -19,7 +19,7 @@ public class CouponTypeCreationRequestDto {
 
     @Min(0)
     @Max(100)
-    private Double discountRate;
+    private Integer discountRate;
 
     @Min(0)
     private Long discountAmount;
@@ -29,7 +29,7 @@ public class CouponTypeCreationRequestDto {
      *
      * @param discountRate 정액처리시 null값이 들어옵니다.
      */
-    public void setDiscountRate(Double discountRate) {
+    public void setDiscountRate(Integer discountRate) {
         this.discountRate = discountRate;
     }
